Finished Problem 2
This was much easier than Problem 1 because I had all of the grunt work typed out already.

I had to nest a for loop in a while loop, probably could be done with a single loop byt I was saving time. Pretty sure there's an exponential formula for this as well.

struggles:
- didn't know the syntax for "True and False" in python
- left some silly things in that probably don't need to be there

learned:
- Apparently I am able to call the month variable even though I am out of the declared scope of the month variable. It is declared in the while loop and I can access it after I leave the while loop